Adds legacy players to game space and uses their svg from the database
