--- src/server/modules/chained-horse/socket/state/players.ts
@@ -1,3 +1,0 @@
-// This file is now deprecated - all state management moved to world.ts
-// Remove this file once the actor system is fully tested
-export {};
