+++ src/utils/getImage.ts
@@ -1,4 +1,5 @@
 import { ActorType } from '../server/types/actor';
+import { decode, encode } from 'js-base64'
 
 export const getImage = (type: ActorType, id?: string): string => {
     switch (type) {
@@ -13,3 +14,16 @@ export const getImage = (type: ActorType, id?: string): string => {
             throw new Error(`Unknown actor type: ${type}`);
     }
 };
+
+export const getSVG = (imgsrc: string) => {
+    const svg = decode(imgsrc.split(',')[1])
+    const param = imgsrc.split(',')[0]
+    const _svg = []
+    svg.split('><').map((bit, i) => bit.indexOf(`rect width='32' height='32' fill='#`) !== -1 ?
+        _svg.push(`<rect width='32' height='32' fill='transparent' />`)
+        : i === 0 ? _svg.push(`${bit}>`)
+        : i === svg.split('><').length - 1 ? _svg.push(`<${bit}`)
+            : _svg.push(`<${bit}>`) // Manky.. strips out the background
+    )
+    return param + ',' + encode(_svg.join(''))
+}
